+++ hello-world/src/ui/utils/device/index.ts
@@ -1,4 +1,0 @@
-export {default as getDeviceType} from "./getDeviceType";
-export {DeviceTypes} from "./device.constants";
-
-export type {DeviceType} from "./device.type";