+++ hello-world/src/ui/utils/index.ts
@@ -1,1 +1,7 @@
+export {default as isServerSideRendering} from "./application/isServerSideRendering";
+
 export * as ClassesUtil from "./classes/classes";
+
+export {DeviceTypes} from "./device/device.constants";
+export type {DeviceType} from "./device/device.type";
+export {default as getDeviceType} from "./device/getDeviceType";